Add Phoenix 6 BETA as vendordep
The Phoenix 6 BETA version has been added as a dependency to this project.
This should allow us to get a start on the development proccess.

IMPORTANT!!

This version of Phoenix is UNSUPPORTED by CTRE with the kickoff release of WPILib and the 2024 game tools.
It is likely that deployments onto a rio will fail. Once a stable release is announced, we will use it.