Update Swerve.java to use AHRS & update vendordeps
Updated Swerve.java to use the NavX gyro while we wait for the purchase
order to go through. The code to utilize a pidgeon2 has been commented out and should be kept.
In addition, the NavX/AHRS library for kauai labs has been added as a dependency to the project.