Clean up LED code
- Separated logic to control multiple LEDs into a separate class
- Modified LEDHandler to handle changing LED color
- Put constants previously in LED.java into separate classes and
  Constants.java

IMPORTANT!!
LED colors must be set. In addition, we are unsure of the amount of LED
strips there will be on the robot, or even if we will be using CANdle's
or not. This code has not been thoroughly reviewed and should be
subject to harsh review and editing for safety, optimization, and
performance.

Co-Authors:
Carlos Nava